Split test function model library from driver so pickle works
